Update runners and function types for actions.
Functions type definitions are updated to add bulk action types, namespacing,
and a new piece of metadata: `modelSelectionField`.

The model selection field allows us to distinguish between the model's API
identifier, and the field in GraphQL that applies to the `select` option given
to our operations. With bulk actions, for example, the model API identifier may
be "widget", but the GraphQL response is served under a field named "widgets".
The model API identifier is still useful for bulk actions, because we can fetch
hydrations from `gadgetMeta` with it.