Ensure connected event always heard by `connectionOpen`.
Previously, there was a chance that if the connection was established very
quickly, we may have missed the "connected" event from the websocket before we
created the listener in `connectionOpen`.

To handle this scenario, we ensure a connected listener is given in the
construction of the subscription client that tracks the status.