Use urql's cache exchange when running in the browser by default
urql, the graphql react client library we use, has a light weight caching feature that reloads queries shown on screen when any mutation that might have changed the data runs. See https://formidable.com/open-source/urql/docs/basics/document-caching/ . You get it out of the box when using urql normally, but right now when using Gadget's react hooks, it is off by accident. This is because we ask users to pass in our highly-configured urql client, and we don't include the exchange. I didn't include it initially because it doesn't make much sense in a server-to-server communication context where you expect and API client to make API calls and not cache data that happened to have been fetched recently, but in the browser, urql has chosen this as default behaviour and I agree it makes sense.

Fixes [sc-3190]