Add support for a custom auth method which can be used to implement platform native auths from userland
To build support for Shopify session token authentication, we need the client to be able to set arbitrary headers to send back to the Gadget backend. We don't wanna build super bespoke shopify code into a client used in a variety of non-shopify places, so this adds a hook were folks constructing the connection can create connections that pass a function that does whatever it wants to the outgoing request. Woop woop!