Switch the Provider to use Shopify Session Tokens for auth instead of a code exchange to get incongnito support.
Fixes GGT-1757 and fixes GGT-1681

We got a report from a user listing their app on the Shopify App Store that their app didn't pass review because it couldn't auth in incognito. After much investigation, we determined that the current localStorage/sessionStorage strategy won't ever work inside an iframe inside an incognito window with the default browser privacy settings. Browsers treat iframe'd localStorage/sessionStorage the same way as third party cookies, which means they just fully disallow access to these objects inside an iframe inside an incognito window. Just touching the value on `window` throws an error.

So, we're gonna do the thing we were intending on doing eventually a bit earlier, and make our provider use Shopify's suggested implementation for third party communication within the iframe context: session tokens. Session tokens are a JWT token we ask the AppBridge for, and then forward along to the Gadget backend, which knows how to decode them and bootstrap a Gadget session to match.  This means we can remove the (painstakingly crafted) code exchange stuff we were doing before that managed a token client side in storage, as we're now just a client of Shopify's JWTs. They're the authority on what a session is, we just need to proxy it through to Gadget.