[hooks] Return a Gadget-owned error wrapper object with easy access to validation errors from GraphQL responses
When a Gadget backend rejects an action because the supplied params are invalid, it returns an error in the format like:
```
{
  success: false,
  errors: {
    message: "GGT_INVALID_RECORD: Record is invalid. 2 error needs to be corrected",
    validationErrors: [
       { field: "title", message: "is too short" },
       { field: "body", message: "is required" }
    ]
}
```

The point of all that is to always return a top level error message regardless of what kind of error happened (be it like a backend error where all we have is a message) and then provide structured error details if we do have them (the by-field error messages).

This data is really useful for driving form validations where you want to show a user which fields are broken. They need to correct those fields before they can move on, so we need to make this data from the GraphQL response available in React land really easily. Before this change, the way you would get at it is like so:

```typescript
const [{data, fetching, error}, mutate] = useAction(api.widget.create);
const validationErrors = error?.graphQLErrors.?[0]?.originalError.validationErrors // any
```

which sucks IMO. The type of the `error` object returned before this change is an `urql` `CombinedError` object which wraps all the errors that can occur client side, like network errors and GraphQL execution errors. I think we should return our own type that works like this:

```typescript
const [{data, fetching, error}, mutate] = useAction(api.widget.create);
const validationErrors = error.validationErrors // FieldErrorMessage[] | undefined
```

This does just that! We introduce a new class called `ErrorWrapper` which is the `error` in all our React hook results, and give it nicer types for having these validation errors on them.