Bump urql in all client packages.
https://github.com/FormidableLabs/urql/blob/main/packages/core/CHANGELOG.md

urql has went through a major version bump. Nothing particularly breaking in
here, minus the requirement to provide a variables object to some of the
methods. The multipart fetch exchange went to 1.0 to support the `@urql/core`
bump.

This is a prerequisite to introducing the same bump in the gadget server, which
in turn is needed for introducing some other urql packages that are being built
around version 3.