hotfix: Return a GadgetRecordList from the useFindMany hook
It's important that react hook users are able to inquire about the pageInfo returned by a connection, so they can paginate forwards or backwards and know if those pages exist! We tried to make this possible previously but failed by accident -- we were still returning a plain old array of records instead of the extended-array GadgetRecordList object. This updates us to definitely return that fancier object to match the docs that already describe using it, and adds a test that actually exercises those bits to prove it is working.