Don't use localStorage or sessionStorage if they aren't supported inside the GadgetConnection
This fixes an issue where if configured to use localStorage via the default browserSession auth mode, Gadget would break in incognito windows. This is a big problem for Shopify apps which are expected to work in incognito windows! Instead, we fall back to the temporary in memory storage function. For shopify apps, we rely on the custom auth method being set when the client is passed to the provider later.