Ensure the react hooks return objects with the same identity if nothing changes on rerender
urql takes great care to ensure that if your component rerenders and calls `useQuery` again, but nothing about the query has changed, you get back the same object. This is so that if you use that object in a dependencies array, it's identity is stable and react won't expire memos or re-run useEffects.

Prior to this change, we didn't keep this same property. If you wrote code like this:

```javascript
const [{ data, fetching, error}] = useFindFirst(api.someModel);
const [something, setSomething] = useState(null)

useEffect(() => {
    setSomething(data)
}, [data]);
```

You'd accidentally get an infinite loop because the useEffect would trigger a rerender because it called setstate, which would rerun our react hook, which would return new data, which would be passed to the array of dependencies for use effect, which would trigger it to rerun, etc.

Instead, let's keep the same property that urql has, which is that you only get a new result object when something about the result changes.