Add an option for not immediately sending GET requests
While working on a form that submitted data to a GET endpoint I found it annoying that `useFetch` insisted on issuing the request as soon as the component renders. I think that is the right default behaviour (kinda like how `useFindMany` starts finding right away), but should be opt-out-able. This adds a new `sendImmediately: false` option to the `useFetch` hook allowing disabling this automatic-send behavior for GET requests.