Update the react provider to accept a full instance of an api client instead of just the urql client
We're likely to need access to more than just the urql client in the
react hooks over time. We already have a nasty hack to get back from the
urql client to the actual GadgetConnection object in order to power
`useFetch`, and I expect more of this kind of thing to happen in the
future. Let's update the API from being

```typescript
<Provider value={api.connection.currentClient} />
```

to be

```typescript
<Provider api={api} />
```

This matches the Shopify provider component, which takes an `api`
prop as well.

We ensure this change is backwards compatible by still supporting the
`value` prop, and still expecting it to be passed an urql client. Added
a few tests to this effect as well.