Detect embeddedness in the Shopify Mobile apps where we're not iframed but instead embedded
The Gadget app bridge wrapper library can power embedded apps that use the App Bridge, or non-embedded apps that don't. In the embedded case, we need to detect if the app bridge is embedded correctly, and if not, redirect to the embedded experience from the outside. We previously only detected if we were embedded using iframe-ness, but, inside the Shopify Mobile iOS and Android apps, the WebView that shows your app *isn't* an iframe -- it uses a different, other transport mechanism to communicate with Shopify. In order to detect embeddedness on those platforms, we look for Shopify's fancy transport instead.

This fixes a redirect loop caused by the Gadget shopify app bridge wrapper library within the mobile apps, where without that detection, we'd be wrong about if we were embedded or not, and try to redirect back to where we were already infinitely.