Add the select option to the internal model manager
We recently added a feature to the internal API for only selecting some
fields of the returned records. This gives improved performance as the
default can exclude computed fields which are often expensive and
unused. This feature already exists in the GraphQL API, but the JS
client didn't know about it. This teaches it!

I also took the opportunity to type the options and improve the JSDoc comments for the
InternalModelManager class so folks know what the heck this thing does.