Ensure the low-level useGadgetQuery and useGadgetMutation hooks don't suspend by default
In #215, we introduced automatic suspense support for the React query hooks, which allows developers (and us) to mark specific queries with "we want them to suspend" so that urql can do its thing under the hood and do so. We wanted this for a nice api in the `useSession` etc hooks.

To implement this though, we need to turn urql's suspense support on at the client level. This is a constructor option which makes suspense available at all *and* makes it the default for all queries. So that this wasn't a breaking change for all users, I turned on suspense support in the client, but then made all queries pass an explicit `suspense: false` down to urql so the behaviour was unchanged between versions. If a query wants to suspend, it has to opt-in by passing `suspense: true`.

I bungled this though, and did it for all the high level hooks like `useFindMany` and `useAction`, but I forgot that we export and make use of the low level `useGadgetQuery` hook for running manual GraphQL queries. Critically, we use this in the `@gadgetinc/react-shopify-app-bridge` package to detect installed status. The effect of this miss is that when the old react-shopify-app-bridge package is paired with a new api-client-core version, the bridge gets a client marked with `suspense: true` on the constructor, and then uses the low level `useGadgetQuery` hook, which doesn't apply the per-query default. The hook then suspends, which we learned breaks shopify installation.