Add support for live queries
Using the @live query directive, we can ask Gadget to give us back a stream of data as it changes from the backend! We can use the graphql-ws transport we already have setup and productionized for this, and we can use urql's ability to rerender at will to trigger a nice rerender. To start, live queries are only available in the React hooks for rerendering.