Add an extra option for route path building for when the endpoint isn't the public URL of the app
Outside of Gadget's internal infrastructure, the endpoint the API client is configured with is the public HTTP URL for the app + "/api/graphql". This means we can use the endpoint to infer the public HTTP URL for the app, which we use for building a fully qualified URL when someone does `api.fetch("/some/relative/path")` or `useFetch("/some/relative/path")`.

Sadly, this breaks inside Gadget's infrastructure, where we set up api clients with a custom GraphQL endpoint that uses our internal sandbox<->api connection for performance. The endpoint inside the sandbox is `api.gadget-services.net`, which means `api.fetch("/foo"` becomes a request to `https://api.gadget-services.net/foo`, which isn't right at all. It's weird that folks would be doing `api.fetch` within the sandbox to make an API call back to the sandbox they are already in, but stranger things have happened, and I think it should still work despite being roundabout.

So, we need a way to build the correct absolute URLs for relative inputs from within the sandbox. To KISS, this just adds a new option that we can pass to the api client in the sandbox. It's backwards and forwards compatible in that old api clients wont break when passed this, and new api clients don't need it to be passed. API clients constructed outside the sandbox by anyone other than us can use the public API endpoint for inference and don't need to pass this option either.