Fix typecheck in all packages after switch to ESM
When I switched js-clients to start publishing both CJS and ESM, I had to mess with the tsconfigs and add separate ones for each build mode. This works fine for the `src` directory of each project, but I missed the fact that because those tsconfigs only apply to the src dirs, running a build only typechecks the `src` dirs. Nothing in CI was typechecking the `spec` dirs or the test bundless, whoops!

This adds an explicit new CI step to typecheck all the files in every package that has a `typecheck` script defined, and fixes a few errors that crept in.