Add support for running Gadget clients in environments without a Websocket implementation
This allows use in places like Shopify's locked down extension environment, where we do have a `globalThis.fetch`, but don't have a `globalThis.WebSocket`. That's ok -- extensions don't really need to open up transactions against Gadget and so don't need websockets. In a subscriptions/livequery-y future that'd change, but investing in a whole other transport is a lot of work. So, let's at least unblock using the client for normal reads and writes!