Fix support for invoking bulk actions with flattened params
Support for invoking bulk actions with flattened params in react was broken, and https://github.com/gadget-inc/js-clients/pull/267 accidentally broke it further.

It was broken since inception because the tests were asserting incorrectly that the produced `input` params weren't nested -- they should have been, oops! Updated the tests and stopped using snapshot tests to describe this so no one is tempted to break it.

Then, to fix the issue and properly disambiguate the params, I had to change the behaviour of the disambiguator a little bit. To preserve the behaviour for the change in #267, we need to respect the idea that if an action does not accept model input, we shouldn't do any nesting of variables. This does this by giving the new `action.acceptsModelInput` value precedence over `action.hasCreateOrUpdateEffect`, which was the ancient way of asking this question. That old way keeps support for old API clients, but the new way is more trustworthy. For things like the signUp action, `action.acceptsModelInput` is false, but `action.hasCreateOrUpdateEffect` is true. So, we stop `||`-ing them, and instead give the new thing precedence. This way, for actions with hardcoded param sets like `user.signUp`, we never nest.

This fixes the over-eager nesting issue for bulk actions as well.