Fix support for POST requests run in strict mode from useFetch
This fixes a very subtle bug in `useFetch`. When running a fetch request with `shouldSendImmediately: false` and `json: true`, users would get an error looking like this:

```
Uncaught (in promise) ErrorWrapper: [GraphQL] The user aborted a request.
    at _ErrorWrapper.forClientSideError (https://fetch-testing--development.gadget.app/node_modules/.vite/deps/@gadgetinc_react.js?v=19130701:484:12)
    at dispatchError (https://fetch-testing--development.gadget.app/node_modules/.vite/deps/@gadgetinc_react.js?v=19130701:966:32)
    at https://fetch-testing--development.gadget.app/node_modules/.vite/deps/@gadgetinc_react.js?v=19130701:1048:25
```

This is because the POST request was immediately, erroneously being aborted as soon as it was run. We don't want to abort the request, we want to run it!

We have a bunch of existing code to abort requests on component unmount for good hygiene. We don't need the response for a component that is no longer on screen and we don't want to charge users (or make expensive openai requests) that take a long time that no one wants anymore. So, we use a useEffect teardown to abort defunct requests. This works good, but, the bug stemmed from the dependencies to that useEffect changing at the wrong time. We were accidentally mutating one of the options objects that gets passed to that useEffect as a dependency on request start, which had the effect of triggering the useEffect to do its unmount/remount thing immediately on the next render. Simply running the request should not ever cause the dependencies to change.

Most confusingly, this mutate-the-options-and-trigger-a-depenendcies-change only happened when you passed a different option in to the hook (`json: true`) which I think is why this bug lurked so long.

The fix is to just not mutate the options object we're using as a dependency in the dependencies array, which means no dependencies change and the useEffect doesn't abort the request. We keep it around though for the actual unmount case.

This also showed a bit of a test hole, which I fixed. useEffects are also pains in the ass for react strict mode which I initially thought was the culprit. I kept those tests around, and added a test that passes in `json` and failed before these changes.