Add live query support to the imperative operation runners
I chose to use async iterators for this api. The end result would look like:

```
for await (const result of api.widget.findOne("123")) {
  console.log(result);
}
```

`result` there would be a fully formed `GadgetRecord` instance, and each time it changed on the backend, we'd push a whole new instance out the async iterator.

[no-changelog-required]