Add support for calling fields with arguments
GraphQL supports fields at arbitrary depths that take arguments. We already support passing args at the base-most level of a query, like with:

```typescript
await api.widget.findMany({filter: { ... }});
```

but, GraphQL supports sending args at any depth at all! For example, you might want to fetch a filtered list of widgets, and their child gizmos, and filter the list of gizmos for each widget as well! In GraphQL, this looks like:

```graphql
query {
 widgets(filter: { inventoryCount: { greaterThan: 10 } }) {
   edges {
     node {
       id
       inventoryCount
       gizmos(first: 10, filter: { published: { equals: true } }) {
         edges {
           node {
             id
             published
           }
        }
      }
    }
  }
}
```

We don't currently support this in the JS client, even though our GraphQL schema supports it fine. The main use case is adjusting how many child records you get back for each parent on a nested fetch, as well as filtering those child records, and asks in discord have cropped up a bunch for this. In this code snippet, where do you put arguments for gizmos?:

```typescript
// before, no calls supported
await api.widget.findMany({
  select: {
    id: true,
    name: true,
    gizmos: {
      edges: {
        node: {
          id: true,
          published: true,
        }
      }
    }
  }
});
```

This adds a thing to do this just this! There's two parts to it: knowing which fields take arguments, which we currently don't really describe, and then adding a JS-land syntax for passing calls at an arbitrary place in the selection. I chose to do this with a new primitive that looks like this:

```typescript
await api.widget.findMany({
  select: {
    id: true,
    name: true,
    gizmos: Call({
      first: 10,
      filter: { published: { equals: true } },
    }, {
      edges: {
        node: {
          id: true,
          published: true,
        }
      }
    })
  }
});
```

[no-changelog-required]