Fix esm/cjs typescript setup use matching values of module and moduleResolution
The are-the-types-wrong tool identified this error with api-client-core: https://github.com/arethetypeswrong/arethetypeswrong.github.io/blob/main/docs/problems/InternalResolutionError.md , stemming from mismatched values of the typescript options for `module` and `moduleResolution`. This makes us use the newer version of `moduleResolution` when doing ESM compilation, which caught a few real bugs with missing file extensions, and got a lot more picky about the types of the packages we depend on. `@n1ru4l/push-pull-async-iterable-iterator` also has busted types internally because it uses the same .d.ts files to describe both CJS and ESM, but we want to keep checking types in our libs, so we patch it to fix it.

This fixes up the typescript configs for the other two packages as well.