Ensure that semantically meaningless changes to dates or belongsto fields aren't reported as such
There's a bug where GadgetRecord reports changes to fields when the object changed identity but not value. In the case of dates, this happens with assigning dates that are the same instant but from different moments in time, and in the case of BelongsTo fields, this happens when setting `record.shopId = { _link: "123" }`, when the prior value was just "123". In these instances, we change our definition of equality to ensure these are considered the same value.

For dates, this was actually already the case, but I think what's happening Gadget side is the value is showing up as a string, and so they are not equal. The bug was real for belongstos though.

Partly fixes GGT-6392