fix(autocomponents): Add submodule resolution package.json files
Are the types wrong helpfully pointed out that older node module
resolution strategies do not understand exports from the package.json
for submdodules, so our `import ... from @gadgetinc/react/auto/polaris`
imports weren't going to work.

This probably doesn't matter except that many people use a tsconfig.json
file with moduleResolution: 'node', which triggers node 10 module
resolution lookups which ignore package.json exports.

In order to make this work, we need a folder in the root folder for each
level of the submodule import, with a packge.json file that points us to
the main entry point for each submodule, so for example, for
`react/auto/polaris`, in the built pacakage, there needs to be a file
`react/auto/polaris/package.json` which points us back to
dist/cjs/auto/polaris/index.js.

This was still reporting a type problem, which we eventually resolved by
adding the folder to the the `files` property of the root package.json
in order to include it in the built package.