Add support for an install state hint passed from the server side to the client that allows the provider to fast-path
In order to lower app LCP, we want to avoid having to do a whole other roundtrip to the server every time the page loads to check the installed state of a shop. We need to do this in some circumstances -- for off-platform apps, or to trigger the install of a shop the very first time a shop using managed installs is seen. So, we can't just remove this request, but, we can avoid blocking on it if the backend says its ok to. The backend already needs to check install state when rendering the page in order to inject the right shopify api key, so we can have it pass a little bit more info about the install state when it does that.