Ensure JSON input errors prevent autoform submission
Previously, the `useJSONInputController` was tracking its own error from invalid JSON inputs in its own `setState` call. This state was local to the component, so it didn't stop the form from being submitted. Instead, we need to store the error in the form state such that `react-hook-form` knows about it, and can revalidate using the validation schema. This implements that! It's still a little tricky because we need to store the raw input string the user is typing into and store the real JSON value in the form state so that it gets submitted. The controller now queues the validator that there has been a parse error by writing a special object indicating that there has been a parse error into the form state.