useActionForm: ensure both shorthand and fully-qualified field names can be passed to send
The new data-masking functionality in `useActionForm` should be compatible with people using either shorthand field keys or longhand field keys for both create and update actions. This wasn't quite the case before for two reasons: we were only telling the data masker to try both keys for update actions, and we were only passing in pre-disambiguated data to try for update actions. This fixes both of those small issues and adds tests!