Add belongs to path suffix for hidden fields setting belongstos
This fixes using hidden inputs against belongs tos. Previously, the hidden input would set the string ID right into the form values as the value at the field, which doesn't match what `useActionForm` is expecting or what the belongs to controller already does. For belongstos, we need to suffix the path with a `.id` to have useActionForm process the form properly.

This reverts commit d5259ed38bef61771ff85d59777ec2f7827e1e2b which I accidentally landed on main prematurely.