Don't use shadcn form -- it overrides react-hook-form context that we need
The Form component we were using from shadcn was not the same as the mainline component and had been doctored, and the mainline one breaks things! We can't use it because it fucks with the react-hook-form state and re-implements part of our stuff. Instead, let's just hardcode our own container for now.