Strengthen internal .deleteMany type to require being passed a filter
A user on discord invoked `api.thing.internal.deleteMany(["1", "2", "3"])` which is wrong -- the function accepts an options object with a filter and a sort, but the type had those properties marked as optional, which meant that the array actually passed the typecheck! The right invocation is `api.thing.internal.deleteMany({filter: { ...}})`. Since this is a really sensitive operation, I think we should be more annoying about the required inputs so that users are forced to pass a filter of some sort. They can still delete everything by doing `api.thing.internal.deleteMany({filter: {}})`, but I think we should make that done explicitly instead of by accident if you don't pass the right options.

This is a breaking types change which I am ok to just YOLO -- functionality is unaffected but I think we can consider it a bugfix.